Move program-specific HQ task start code to a separate file
